Revert "stockfish.wasm: Do not create worker for fake thread"
This reverts commit 702fb2557f34bfa815ee89dbd8216bd2f79ba595.

+++ src/thread.h
@@ -44,12 +44,12 @@ class Thread {
 
   Mutex mutex;
   ConditionVariable cv;
-  int idx;
+  size_t idx;
   bool exit = false, searching = true; // Set before starting std::thread
   pthread_t nativeThread;
 
 public:
-  explicit Thread(int);
+  explicit Thread(size_t);
   virtual ~Thread();
   virtual void search();
   void clear();
